@@ -1,6 +1,7 @@
 library firebase_auth;
 
 import 'dart:async';
+import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:math';
@@ -14,8 +15,8 @@ import 'package:built_value/serializer.dart';
 import 'package:firebase_auth/src/models/serializers.dart';
 import 'package:firebase_common/firebase_common.dart';
 import 'package:firebase_internal/firebase_internal.dart';
-import 'package:googleapis/identitytoolkit/v3.dart' as gitkit;
 import 'package:googleapis/identitytoolkit/v3.dart';
+import 'package:googleapis/identitytoolkit/v3.dart' as gitkit;
 import 'package:hive/hive.dart';
 import 'package:http/http.dart';
 import 'package:meta/meta.dart';
@@ -31,16 +32,19 @@ part 'src/data/secure_token_service.dart';
 part 'src/data/user_storage.dart';
 part 'src/firebase_auth.dart';
 part 'src/firebase_user.dart';
+part 'src/models/action_code_info.dart';
 part 'src/models/action_code_settings.dart';
 part 'src/models/auth_operation_type.dart';
-part 'src/models/credentials/auth_credential.dart';
-part 'src/models/credentials/provider_type.dart';
+part 'src/models/auth_credential.dart';
+part 'src/models/provider_type.dart';
 part 'src/models/impl.dart';
 part 'src/models/oob_code_type.dart';
 part 'src/models/secure_token.dart';
 part 'src/models/sign_in_with_game_center_request.dart';
 part 'src/models/user.dart';
+part 'src/models/user_update_info.dart';
 part 'src/util/api_key_client.dart';
 part 'src/util/errors.dart';
 part 'src/util/http_server.dart';
+part 'src/util/sequential_runner.dart';
 part 'src/util/web_utils.dart';
